+++ apps/www/pages/index.tsx
@@ -13,6 +13,7 @@ import PageLayout, { Props as PageLayoutProps } from "~/pages/PageLayout"
 import createListStaticPropsGetter from "~/ssg/createListStaticPropsGetter"
 import CompressedSWRConfig from "~/swr/CompressedSWRConfig"
 import HeaderNav from "~/ui/HeaderNav"
+import QuickLinks from "~/ui/QuickLinks"
 import SiteTitle from "~/ui/SiteTitle"
 import ContributionCTAView from "~/views/ContributionCTAView"
 import ImageListView from "~/views/ImageListView"
@@ -110,8 +111,17 @@ const PageComponent: NextPage<Props> = ({ fallback, ...props }) => (
                 </PaginationContainer>
             </section>
             <ContributionCTAView />
+            <section>
+                <header>
+                    <h2>Quick Links</h2>
+                    <p>
+                        The most popular taxon pages on <SiteTitle />, as determined by actual traffic:
+                    </p>
+                </header>
+                <QuickLinks />
+            </section>
             <section style={{ minWidth: "100vw" }}>
-                <header style={{ maxWidth: "11in", margin: "0 auto" }}>
+                <header style={{ maxWidth: "11in", margin: "0 auto", paddingLeft: "1rem", paddingRight: "1rem" }}>
                     <HeaderNav
                         buttons={[
                             {