+++ apps/www/src/ui/QuickLinks/QUICK_LINKS.ts
@@ -1,0 +1,262 @@
+import { QuickLinkNode } from "./QuickLinkNode"
+
+const QUICK_LINKS: QuickLinkNode = {
+    label: "life",
+    slug: "pan-biota",
+    uuid: "8f901db5-84c1-4dc0-93ba-2300eeddf4ab",
+    children: [
+        {
+            label: "bacteria",
+            slug: "eubacteria",
+            uuid: "--",
+        },
+        {
+            slug: "eukaryota",
+            uuid: "--",
+            children: [
+                {
+                    label: "plants",
+                    slug: "archaeplastida",
+                    uuid: "--",
+                    children: [
+                        {
+                            label: "grasses",
+                            slug: "gramineae",
+                            uuid: "--",
+                        },
+                        {
+                            label: "oaks",
+                            slug: "quercus",
+                            uuid: "--",
+                        },
+                    ],
+                },
+                {
+                    slug: "opisthokonta",
+                    uuid: "--",
+                    children: [
+                        {
+                            label: "fungi",
+                            slug: "fungi",
+                            uuid: "--",
+                        },
+                        {
+                            label: "animals",
+                            slug: "metazoa",
+                            uuid: "--",
+                            children: [
+                                {
+                                    label: "sponges",
+                                    slug: "porifera",
+                                    uuid: "--",
+                                },
+                                {
+                                    slug: "nephrozoa",
+                                    uuid: "--",
+                                    children: [
+                                        {
+                                            slug: "protostomia",
+                                            uuid: "--",
+                                            children: [
+                                                {
+                                                    slug: "trochozoa",
+                                                    uuid: "--",
+                                                    children: [
+                                                        {
+                                                            label: "annelids",
+                                                            slug: "annelida",
+                                                            uuid: "--",
+                                                        },
+                                                        {
+                                                            label: "mollusks",
+                                                            slug: "Mollusca",
+                                                            uuid: "--",
+                                                            children: [
+                                                                {
+                                                                    label: "snails",
+                                                                    slug: "gastropoda",
+                                                                    uuid: "--",
+                                                                },
+                                                            ],
+                                                        },
+                                                    ],
+                                                },
+                                                {
+                                                    slug: "arthropoda",
+                                                    uuid: "--",
+                                                    children: [
+                                                        {
+                                                            label: "spiders",
+                                                            slug: "araneae",
+                                                            uuid: "--",
+                                                        },
+                                                        {
+                                                            label: "insects",
+                                                            slug: "insecta",
+                                                            uuid: "--",
+                                                            children: [
+                                                                {
+                                                                    label: "damselflies & dragonflies",
+                                                                    slug: "odonata",
+                                                                    uuid: "--",
+                                                                },
+                                                                {
+                                                                    slug: "endopterygota",
+                                                                    uuid: "--",
+                                                                    children: [
+                                                                        {
+                                                                            label: "bees",
+                                                                            slug: "apiformes",
+                                                                            uuid: "--",
+                                                                        },
+                                                                        {
+                                                                            slug: "aparaglossata",
+                                                                            uuid: "--",
+                                                                            children: [
+                                                                                {
+                                                                                    label: "beetles",
+                                                                                    slug: "coleoptera",
+                                                                                    uuid: "--",
+                                                                                },
+                                                                                {
+                                                                                    slug: "panorpida",
+                                                                                    uuid: "--",
+                                                                                    children: [
+                                                                                        {
+                                                                                            label: "flies",
+                                                                                            slug: "diptera",
+                                                                                            uuid: "--",
+                                                                                        },
+                                                                                        {
+                                                                                            label: "moths",
+                                                                                            slug: "lepidoptera",
+                                                                                            uuid: "--",
+                                                                                        },
+                                                                                    ],
+                                                                                },
+                                                                            ],
+                                                                        },
+                                                                    ],
+                                                                },
+                                                            ],
+                                                        },
+                                                    ],
+                                                },
+                                            ],
+                                        },
+                                        {
+                                            label: "vertebrates",
+                                            slug: "vertebrata",
+                                            uuid: "--",
+                                            children: [
+                                                {
+                                                    label: "amphibians",
+                                                    slug: "amphibia",
+                                                    uuid: "--",
+                                                },
+                                                {
+                                                    slug: "amniota",
+                                                    uuid: "--",
+                                                    children: [
+                                                        {
+                                                            slug: "reptilia",
+                                                            uuid: "--",
+                                                            children: [
+                                                                {
+                                                                    label: "lizards",
+                                                                    slug: "squamata",
+                                                                    uuid: "--",
+                                                                },
+                                                                {
+                                                                    slug: "archelosauria",
+                                                                    uuid: "--",
+                                                                    children: [
+                                                                        {
+                                                                            label: "turtles",
+                                                                            slug: "testudines",
+                                                                            uuid: "--",
+                                                                        },
+                                                                        {
+                                                                            label: "dinosaurs",
+                                                                            slug: "dinosauria",
+                                                                            uuid: "--",
+                                                                            children: [
+                                                                                {
+                                                                                    label: "birds",
+                                                                                    slug: "aves",
+                                                                                    uuid: "--",
+                                                                                    children: [
+                                                                                        {
+                                                                                            label: "owls",
+                                                                                            slug: "strigiformes",
+                                                                                            uuid: "--",
+                                                                                        },
+                                                                                    ],
+                                                                                },
+                                                                            ],
+                                                                        },
+                                                                    ],
+                                                                },
+                                                            ],
+                                                        },
+                                                        {
+                                                            label: "mammals",
+                                                            slug: "mammalia",
+                                                            uuid: "--",
+                                                            children: [
+                                                                {
+                                                                    slug: "scrotifera",
+                                                                    uuid: "--",
+                                                                    children: [
+                                                                        {
+                                                                            label: "bats",
+                                                                            slug: "chiroptera",
+                                                                            uuid: "--",
+                                                                        },
+                                                                        {
+                                                                            label: "whales",
+                                                                            slug: "cetacea",
+                                                                            uuid: "--",
+                                                                        },
+                                                                    ],
+                                                                },
+                                                                {
+                                                                    slug: "euarchontoglires",
+                                                                    uuid: "--",
+                                                                    children: [
+                                                                        {
+                                                                            label: "humans",
+                                                                            slug: "homo-sapiens",
+                                                                            uuid: "--",
+                                                                        },
+                                                                        {
+                                                                            label: "rodents",
+                                                                            slug: "rodentia",
+                                                                            uuid: "--",
+                                                                            children: [
+                                                                                {
+                                                                                    label: "mice & rats",
+                                                                                    slug: "muridae",
+                                                                                    uuid: "--",
+                                                                                },
+                                                                            ],
+                                                                        },
+                                                                    ],
+                                                                },
+                                                            ],
+                                                        },
+                                                    ],
+                                                },
+                                            ],
+                                        },
+                                    ],
+                                },
+                            ],
+                        },
+                    ],
+                },
+            ],
+        },
+    ],
+}
+export default QUICK_LINKS